chore: Rename file to sound less confusing with `core/lint.js`
